+++ nematus/initializers.py
@@ -1,0 +1,24 @@
+'''
+Parameter initializers
+'''
+
+import numpy
+
+import theano
+import theano.tensor as tensor
+from theano.sandbox.rng_mrg import MRG_RandomStreams as RandomStreams
+
+def ortho_weight(ndim):
+    W = numpy.random.randn(ndim, ndim)
+    u, s, v = numpy.linalg.svd(W)
+    return u.astype('float32')
+
+def norm_weight(nin, nout=None, scale=0.01, ortho=True):
+    if nout is None:
+        nout = nin
+    if nout == nin and ortho:
+        W = ortho_weight(nin)
+    else:
+        W = scale * numpy.random.randn(nin, nout)
+    return W.astype('float32')
+
