+++ nematus/nmt.py
@@ -20,165 +20,18 @@ from subprocess import Popen
 
 from collections import OrderedDict
 
-from data_iterator import TextIterator
-from util import load_dict
-from alignment_util import *
-
-from domain_interpolation_data_iterator import DomainInterpolatorTextIterator
-
 profile = False
 
+from data_iterator import TextIterator
+from util import *
+from theano_util import *
+from alignment_util import *
 
-# push parameters to Theano shared variables
-def zipp(params, tparams):
-    for kk, vv in params.iteritems():
-        tparams[kk].set_value(vv)
+from layers import *
+from initializers import *
+from optimizers import *
 
-
-# pull parameters from Theano shared variables
-def unzip(zipped):
-    new_params = OrderedDict()
-    for kk, vv in zipped.iteritems():
-        new_params[kk] = vv.get_value()
-    return new_params
-
-
-# get the list of parameters: Note that tparams must be OrderedDict
-def itemlist(tparams):
-    return [vv for kk, vv in tparams.iteritems()]
-
-
-# dropout
-def dropout_layer(state_before, use_noise, trng):
-    proj = tensor.switch(
-        use_noise,
-        state_before * trng.binomial(state_before.shape, p=0.5, n=1,
-                                     dtype=state_before.dtype),
-        state_before * 0.5)
-    return proj
-
-
-# dropout that will be re-used at different time steps
-def shared_dropout_layer(shape, use_noise, trng, value):
-    proj = tensor.switch(
-        use_noise,
-        trng.binomial(shape, p=value, n=1,
-                                     dtype='float32'),
-        theano.shared(numpy.float32(value)))
-    return proj
-
-# make prefix-appended name
-def _p(pp, name):
-    return '%s_%s' % (pp, name)
-
-
-# initialize Theano shared variables according to the initial parameters
-def init_tparams(params):
-    tparams = OrderedDict()
-    for kk, pp in params.iteritems():
-        tparams[kk] = theano.shared(params[kk], name=kk)
-    return tparams
-
-
-# load parameters
-def load_params(path, params):
-    pp = numpy.load(path)
-    for kk, vv in params.iteritems():
-        if kk not in pp:
-            warnings.warn('%s is not in the archive' % kk)
-            continue
-        params[kk] = pp[kk]
-
-    return params
-
-# layers: 'name': ('parameter initializer', 'feedforward')
-layers = {'ff': ('param_init_fflayer', 'fflayer'),
-          'gru': ('param_init_gru', 'gru_layer'),
-          'gru_cond': ('param_init_gru_cond', 'gru_cond_layer'),
-          }
-
-
-def get_layer(name):
-    fns = layers[name]
-    return (eval(fns[0]), eval(fns[1]))
-
-
-# some utilities
-def ortho_weight(ndim):
-    W = numpy.random.randn(ndim, ndim)
-    u, s, v = numpy.linalg.svd(W)
-    return u.astype('float32')
-
-
-def norm_weight(nin, nout=None, scale=0.01, ortho=True):
-    if nout is None:
-        nout = nin
-    if nout == nin and ortho:
-        W = ortho_weight(nin)
-    else:
-        W = scale * numpy.random.randn(nin, nout)
-    return W.astype('float32')
-
-
-def tanh(x):
-    return tensor.tanh(x)
-
-
-def linear(x):
-    return x
-
-
-def concatenate(tensor_list, axis=0):
-    """
-    Alternative implementation of `theano.tensor.concatenate`.
-    This function does exactly the same thing, but contrary to Theano's own
-    implementation, the gradient is implemented on the GPU.
-    Backpropagating through `theano.tensor.concatenate` yields slowdowns
-    because the inverse operation (splitting) needs to be done on the CPU.
-    This implementation does not have that problem.
-    :usage:
-        >>> x, y = theano.tensor.matrices('x', 'y')
-        >>> c = concatenate([x, y], axis=1)
-    :parameters:
-        - tensor_list : list
-            list of Theano tensor expressions that should be concatenated.
-        - axis : int
-            the tensors will be joined along this axis.
-    :returns:
-        - out : tensor
-            the concatenated tensor expression.
-    """
-    concat_size = sum(tt.shape[axis] for tt in tensor_list)
-
-    output_shape = ()
-    for k in range(axis):
-        output_shape += (tensor_list[0].shape[k],)
-    output_shape += (concat_size,)
-    for k in range(axis + 1, tensor_list[0].ndim):
-        output_shape += (tensor_list[0].shape[k],)
-
-    out = tensor.zeros(output_shape)
-    offset = 0
-    for tt in tensor_list:
-        indices = ()
-        for k in range(axis):
-            indices += (slice(None),)
-        indices += (slice(offset, offset + tt.shape[axis]),)
-        for k in range(axis + 1, tensor_list[0].ndim):
-            indices += (slice(None),)
-
-        out = tensor.set_subtensor(out[indices], tt)
-        offset += tt.shape[axis]
-
-    return out
-
-# return name of word embedding for factor i
-# special handling of factor 0 for backward compatibility
-def embedding_name(i):
-    if i == 0:
-        return 'Wemb'
-    else:
-        return 'Wemb'+str(i)
+from domain_interpolation_data_iterator import DomainInterpolatorTextIterator
 
 # batch preparation
 def prepare_data(seqs_x, seqs_y, maxlen=None, n_words_src=30000,
@@ -224,327 +77,6 @@ def prepare_data(seqs_x, seqs_y, maxlen=None, n_words_src=30000,
     return x, x_mask, y, y_mask
 
 
-# feedforward layer: affine transformation + point-wise nonlinearity
-def param_init_fflayer(options, params, prefix='ff', nin=None, nout=None,
-                       ortho=True):
-    if nin is None:
-        nin = options['dim_proj']
-    if nout is None:
-        nout = options['dim_proj']
-    params[_p(prefix, 'W')] = norm_weight(nin, nout, scale=0.01, ortho=ortho)
-    params[_p(prefix, 'b')] = numpy.zeros((nout,)).astype('float32')
-
-    return params
-
-
-def fflayer(tparams, state_below, options, prefix='rconv',
-            activ='lambda x: tensor.tanh(x)', **kwargs):
-    return eval(activ)(
-        tensor.dot(state_below, tparams[_p(prefix, 'W')]) +
-        tparams[_p(prefix, 'b')])
-
-
-# GRU layer
-def param_init_gru(options, params, prefix='gru', nin=None, dim=None):
-    if nin is None:
-        nin = options['dim_proj']
-    if dim is None:
-        dim = options['dim_proj']
-
-    # embedding to gates transformation weights, biases
-    W = numpy.concatenate([norm_weight(nin, dim),
-                           norm_weight(nin, dim)], axis=1)
-    params[_p(prefix, 'W')] = W
-    params[_p(prefix, 'b')] = numpy.zeros((2 * dim,)).astype('float32')
-
-    # recurrent transformation weights for gates
-    U = numpy.concatenate([ortho_weight(dim),
-                           ortho_weight(dim)], axis=1)
-    params[_p(prefix, 'U')] = U
-
-    # embedding to hidden state proposal weights, biases
-    Wx = norm_weight(nin, dim)
-    params[_p(prefix, 'Wx')] = Wx
-    params[_p(prefix, 'bx')] = numpy.zeros((dim,)).astype('float32')
-
-    # recurrent transformation weights for hidden state proposal
-    Ux = ortho_weight(dim)
-    params[_p(prefix, 'Ux')] = Ux
-
-    return params
-
-
-def gru_layer(tparams, state_below, options, prefix='gru', mask=None,
-              emb_dropout=None,
-              rec_dropout=None,
-              **kwargs):
-    nsteps = state_below.shape[0]
-    if state_below.ndim == 3:
-        n_samples = state_below.shape[1]
-    else:
-        n_samples = 1
-
-    dim = tparams[_p(prefix, 'Ux')].shape[1]
-
-    if mask is None:
-        mask = tensor.alloc(1., state_below.shape[0], 1)
-
-    # utility function to slice a tensor
-    def _slice(_x, n, dim):
-        if _x.ndim == 3:
-            return _x[:, :, n*dim:(n+1)*dim]
-        return _x[:, n*dim:(n+1)*dim]
-
-    # state_below is the input word embeddings
-    # input to the gates, concatenated
-    state_below_ = tensor.dot(state_below*emb_dropout[0], tparams[_p(prefix, 'W')]) + \
-        tparams[_p(prefix, 'b')]
-    # input to compute the hidden state proposal
-    state_belowx = tensor.dot(state_below*emb_dropout[1], tparams[_p(prefix, 'Wx')]) + \
-        tparams[_p(prefix, 'bx')]
-
-    # step function to be used by scan
-    # arguments    | sequences |outputs-info| non-seqs
-    def _step_slice(m_, x_, xx_, h_, U, Ux, rec_dropout):
-
-        preact = tensor.dot(h_*rec_dropout[0], U)
-        preact += x_
-
-        # reset and update gates
-        r = tensor.nnet.sigmoid(_slice(preact, 0, dim))
-        u = tensor.nnet.sigmoid(_slice(preact, 1, dim))
-
-        # compute the hidden state proposal
-        preactx = tensor.dot(h_*rec_dropout[1], Ux)
-        preactx = preactx * r
-        preactx = preactx + xx_
-
-        # hidden state proposal
-        h = tensor.tanh(preactx)
-
-        # leaky integrate and obtain next hidden state
-        h = u * h_ + (1. - u) * h
-        h = m_[:, None] * h + (1. - m_)[:, None] * h_
-
-        return h
-
-    # prepare scan arguments
-    seqs = [mask, state_below_, state_belowx]
-    init_states = [tensor.alloc(0., n_samples, dim)]
-    _step = _step_slice
-    shared_vars = [tparams[_p(prefix, 'U')],
-                   tparams[_p(prefix, 'Ux')],
-                   rec_dropout]
-
-    rval, updates = theano.scan(_step,
-                                sequences=seqs,
-                                outputs_info=init_states,
-                                non_sequences=shared_vars,
-                                name=_p(prefix, '_layers'),
-                                n_steps=nsteps,
-                                profile=profile,
-                                strict=True)
-    rval = [rval]
-    return rval
-
-
-# Conditional GRU layer with Attention
-def param_init_gru_cond(options, params, prefix='gru_cond',
-                        nin=None, dim=None, dimctx=None,
-                        nin_nonlin=None, dim_nonlin=None):
-    if nin is None:
-        nin = options['dim']
-    if dim is None:
-        dim = options['dim']
-    if dimctx is None:
-        dimctx = options['dim']
-    if nin_nonlin is None:
-        nin_nonlin = nin
-    if dim_nonlin is None:
-        dim_nonlin = dim
-
-    W = numpy.concatenate([norm_weight(nin, dim),
-                           norm_weight(nin, dim)], axis=1)
-    params[_p(prefix, 'W')] = W
-    params[_p(prefix, 'b')] = numpy.zeros((2 * dim,)).astype('float32')
-    U = numpy.concatenate([ortho_weight(dim_nonlin),
-                           ortho_weight(dim_nonlin)], axis=1)
-    params[_p(prefix, 'U')] = U
-
-    Wx = norm_weight(nin_nonlin, dim_nonlin)
-    params[_p(prefix, 'Wx')] = Wx
-    Ux = ortho_weight(dim_nonlin)
-    params[_p(prefix, 'Ux')] = Ux
-    params[_p(prefix, 'bx')] = numpy.zeros((dim_nonlin,)).astype('float32')
-
-    U_nl = numpy.concatenate([ortho_weight(dim_nonlin),
-                              ortho_weight(dim_nonlin)], axis=1)
-    params[_p(prefix, 'U_nl')] = U_nl
-    params[_p(prefix, 'b_nl')] = numpy.zeros((2 * dim_nonlin,)).astype('float32')
-
-    Ux_nl = ortho_weight(dim_nonlin)
-    params[_p(prefix, 'Ux_nl')] = Ux_nl
-    params[_p(prefix, 'bx_nl')] = numpy.zeros((dim_nonlin,)).astype('float32')
-
-    # context to LSTM
-    Wc = norm_weight(dimctx, dim*2)
-    params[_p(prefix, 'Wc')] = Wc
-
-    Wcx = norm_weight(dimctx, dim)
-    params[_p(prefix, 'Wcx')] = Wcx
-
-    # attention: combined -> hidden
-    W_comb_att = norm_weight(dim, dimctx)
-    params[_p(prefix, 'W_comb_att')] = W_comb_att
-
-    # attention: context -> hidden
-    Wc_att = norm_weight(dimctx)
-    params[_p(prefix, 'Wc_att')] = Wc_att
-
-    # attention: hidden bias
-    b_att = numpy.zeros((dimctx,)).astype('float32')
-    params[_p(prefix, 'b_att')] = b_att
-
-    # attention:
-    U_att = norm_weight(dimctx, 1)
-    params[_p(prefix, 'U_att')] = U_att
-    c_att = numpy.zeros((1,)).astype('float32')
-    params[_p(prefix, 'c_tt')] = c_att
-
-    return params
-
-
-def gru_cond_layer(tparams, state_below, options, prefix='gru',
-                   mask=None, context=None, one_step=False,
-                   init_memory=None, init_state=None,
-                   context_mask=None, emb_dropout=None,
-                   rec_dropout=None, ctx_dropout=None,
-                   **kwargs):
-
-    assert context, 'Context must be provided'
-
-    if one_step:
-        assert init_state, 'previous state must be provided'
-
-    nsteps = state_below.shape[0]
-    if state_below.ndim == 3:
-        n_samples = state_below.shape[1]
-    else:
-        n_samples = 1
-
-    # mask
-    if mask is None:
-        mask = tensor.alloc(1., state_below.shape[0], 1)
-
-    dim = tparams[_p(prefix, 'Wcx')].shape[1]
-
-    # initial/previous state
-    if init_state is None:
-        init_state = tensor.alloc(0., n_samples, dim)
-
-    # projected context
-    assert context.ndim == 3, \
-        'Context must be 3-d: #annotation x #sample x dim'
-    pctx_ = tensor.dot(context*ctx_dropout[0], tparams[_p(prefix, 'Wc_att')]) +\
-        tparams[_p(prefix, 'b_att')]
-
-    def _slice(_x, n, dim):
-        if _x.ndim == 3:
-            return _x[:, :, n*dim:(n+1)*dim]
-        return _x[:, n*dim:(n+1)*dim]
-
-    # projected x
-    state_belowx = tensor.dot(state_below*emb_dropout[0], tparams[_p(prefix, 'Wx')]) +\
-        tparams[_p(prefix, 'bx')]
-    state_below_ = tensor.dot(state_below*emb_dropout[1], tparams[_p(prefix, 'W')]) +\
-        tparams[_p(prefix, 'b')]
-
-    def _step_slice(m_, x_, xx_, h_, ctx_, alpha_, pctx_, cc_, rec_dropout, ctx_dropout,
-                    U, Wc, W_comb_att, U_att, c_tt, Ux, Wcx,
-                    U_nl, Ux_nl, b_nl, bx_nl):
-
-        preact1 = tensor.dot(h_*rec_dropout[0], U)
-        preact1 += x_
-        preact1 = tensor.nnet.sigmoid(preact1)
-
-        r1 = _slice(preact1, 0, dim)
-        u1 = _slice(preact1, 1, dim)
-
-        preactx1 = tensor.dot(h_*rec_dropout[1], Ux)
-        preactx1 *= r1
-        preactx1 += xx_
-
-        h1 = tensor.tanh(preactx1)
-
-        h1 = u1 * h_ + (1. - u1) * h1
-        h1 = m_[:, None] * h1 + (1. - m_)[:, None] * h_
-
-        # attention
-        pstate_ = tensor.dot(h1*rec_dropout[2], W_comb_att)
-        pctx__ = pctx_ + pstate_[None, :, :]
-        #pctx__ += xc_
-        pctx__ = tensor.tanh(pctx__)
-        alpha = tensor.dot(pctx__*ctx_dropout[1], U_att)+c_tt
-        alpha = alpha.reshape([alpha.shape[0], alpha.shape[1]])
-        alpha = tensor.exp(alpha - alpha.max(0, keepdims=True))
-        if context_mask:
-            alpha = alpha * context_mask
-        alpha = alpha / alpha.sum(0, keepdims=True)
-        ctx_ = (cc_ * alpha[:, :, None]).sum(0)  # current context
-
-        preact2 = tensor.dot(h1*rec_dropout[3], U_nl)+b_nl
-        preact2 += tensor.dot(ctx_*ctx_dropout[2], Wc)
-        preact2 = tensor.nnet.sigmoid(preact2)
-
-        r2 = _slice(preact2, 0, dim)
-        u2 = _slice(preact2, 1, dim)
-
-        preactx2 = tensor.dot(h1*rec_dropout[4], Ux_nl)+bx_nl
-        preactx2 *= r2
-        preactx2 += tensor.dot(ctx_*ctx_dropout[3], Wcx)
-
-        h2 = tensor.tanh(preactx2)
-
-        h2 = u2 * h1 + (1. - u2) * h2
-        h2 = m_[:, None] * h2 + (1. - m_)[:, None] * h1
-
-        return h2, ctx_, alpha.T  # pstate_, preact, preactx, r, u
-
-    seqs = [mask, state_below_, state_belowx]
-    #seqs = [mask, state_below_, state_belowx, state_belowc]
-    _step = _step_slice
-
-    shared_vars = [tparams[_p(prefix, 'U')],
-                   tparams[_p(prefix, 'Wc')],
-                   tparams[_p(prefix, 'W_comb_att')],
-                   tparams[_p(prefix, 'U_att')],
-                   tparams[_p(prefix, 'c_tt')],
-                   tparams[_p(prefix, 'Ux')],
-                   tparams[_p(prefix, 'Wcx')],
-                   tparams[_p(prefix, 'U_nl')],
-                   tparams[_p(prefix, 'Ux_nl')],
-                   tparams[_p(prefix, 'b_nl')],
-                   tparams[_p(prefix, 'bx_nl')]]
-
-    if one_step:
-        rval = _step(*(seqs + [init_state, None, None, pctx_, context, rec_dropout, ctx_dropout] +
-                       shared_vars))
-    else:
-        rval, updates = theano.scan(_step,
-                                    sequences=seqs,
-                                    outputs_info=[init_state,
-                                                  tensor.alloc(0., n_samples,
-                                                               context.shape[2]),
-                                                  tensor.alloc(0., n_samples,
-                                                               context.shape[0])],
-                                    non_sequences=[pctx_, context, rec_dropout, ctx_dropout]+shared_vars,
-                                    name=_p(prefix, '_layers'),
-                                    n_steps=nsteps,
-                                    profile=profile,
-                                    strict=True)
-    return rval
-
-
 # initialize all parameters
 def init_params(options):
     params = OrderedDict()
@@ -556,36 +88,36 @@ def init_params(options):
     params['Wemb_dec'] = norm_weight(options['n_words'], options['dim_word'])
 
     # encoder: bidirectional RNN
-    params = get_layer(options['encoder'])[0](options, params,
+    params = get_layer_param(options['encoder'])(options, params,
                                               prefix='encoder',
                                               nin=options['dim_word'],
                                               dim=options['dim'])
-    params = get_layer(options['encoder'])[0](options, params,
+    params = get_layer_param(options['encoder'])(options, params,
                                               prefix='encoder_r',
                                               nin=options['dim_word'],
                                               dim=options['dim'])
     ctxdim = 2 * options['dim']
 
     # init_state, init_cell
-    params = get_layer('ff')[0](options, params, prefix='ff_state',
+    params = get_layer_param('ff')(options, params, prefix='ff_state',
                                 nin=ctxdim, nout=options['dim'])
     # decoder
-    params = get_layer(options['decoder'])[0](options, params,
+    params = get_layer_param(options['decoder'])(options, params,
                                               prefix='decoder',
                                               nin=options['dim_word'],
                                               dim=options['dim'],
                                               dimctx=ctxdim)
     # readout
-    params = get_layer('ff')[0](options, params, prefix='ff_logit_lstm',
+    params = get_layer_param('ff')(options, params, prefix='ff_logit_lstm',
                                 nin=options['dim'], nout=options['dim_word'],
                                 ortho=False)
-    params = get_layer('ff')[0](options, params, prefix='ff_logit_prev',
+    params = get_layer_param('ff')(options, params, prefix='ff_logit_prev',
                                 nin=options['dim_word'],
                                 nout=options['dim_word'], ortho=False)
-    params = get_layer('ff')[0](options, params, prefix='ff_logit_ctx',
+    params = get_layer_param('ff')(options, params, prefix='ff_logit_ctx',
                                 nin=ctxdim, nout=options['dim_word'],
                                 ortho=False)
-    params = get_layer('ff')[0](options, params, prefix='ff_logit',
+    params = get_layer_param('ff')(options, params, prefix='ff_logit',
                                 nin=options['dim_word'],
                                 nout=options['n_words'])
 
@@ -651,11 +183,12 @@ def build_model(tparams, options):
     if options['use_dropout']:
         emb *= source_dropout
 
-    proj = get_layer(options['encoder'])[1](tparams, emb, options,
+    proj = get_layer_constr(options['encoder'])(tparams, emb, options,
                                             prefix='encoder',
                                             mask=x_mask,
                                             emb_dropout=emb_dropout, 
-                                            rec_dropout=rec_dropout)
+                                            rec_dropout=rec_dropout,
+                                            profile=profile)
     
 
     # word embedding for backward rnn (source)
@@ -667,11 +200,12 @@ def build_model(tparams, options):
     if options['use_dropout']:
         embr *= source_dropout[::-1]
 
-    projr = get_layer(options['encoder'])[1](tparams, embr, options,
+    projr = get_layer_constr(options['encoder'])(tparams, embr, options,
                                              prefix='encoder_r',
                                              mask=xr_mask,
                                              emb_dropout=emb_dropout_r,
-                                             rec_dropout=rec_dropout_r)
+                                             rec_dropout=rec_dropout_r,
+                                             profile=profile)
 
     # context will be the concatenation of forward and backward rnns
     ctx = concatenate([proj[0], projr[0][::-1]], axis=proj[0].ndim-1)
@@ -686,7 +220,7 @@ def build_model(tparams, options):
         ctx_mean *= shared_dropout_layer((n_samples, 2*options['dim']), use_noise, trng, retain_probability_hidden)
 
     # initial decoder state
-    init_state = get_layer('ff')[1](tparams, ctx_mean, options,
+    init_state = get_layer_constr('ff')(tparams, ctx_mean, options,
                                     prefix='ff_state', activ='tanh')
 
     # word embedding (target), we will shift the target sequence one time step
@@ -704,7 +238,7 @@ def build_model(tparams, options):
         emb *= target_dropout
 
     # decoder - pass through the decoder conditional gru with attention
-    proj = get_layer(options['decoder'])[1](tparams, emb, options,
+    proj = get_layer_constr(options['decoder'])(tparams, emb, options,
                                             prefix='decoder',
                                             mask=y_mask, context=ctx,
                                             context_mask=x_mask,
@@ -712,7 +246,8 @@ def build_model(tparams, options):
                                             init_state=init_state,
                                             emb_dropout=emb_dropout_d,
                                             ctx_dropout=ctx_dropout_d,
-                                            rec_dropout=rec_dropout_d)
+                                            rec_dropout=rec_dropout_d,
+                                            profile=profile)
     # hidden states of the decoder gru
     proj_h = proj[0]
 
@@ -728,18 +263,18 @@ def build_model(tparams, options):
     opt_ret['dec_alphas'] = proj[2]
 
     # compute word probabilities
-    logit_lstm = get_layer('ff')[1](tparams, proj_h, options,
+    logit_lstm = get_layer_constr('ff')(tparams, proj_h, options,
                                     prefix='ff_logit_lstm', activ='linear')
-    logit_prev = get_layer('ff')[1](tparams, emb, options,
+    logit_prev = get_layer_constr('ff')(tparams, emb, options,
                                     prefix='ff_logit_prev', activ='linear')
-    logit_ctx = get_layer('ff')[1](tparams, ctxs, options,
+    logit_ctx = get_layer_constr('ff')(tparams, ctxs, options,
                                    prefix='ff_logit_ctx', activ='linear')
     logit = tensor.tanh(logit_lstm+logit_prev+logit_ctx)
 
     if options['use_dropout']:
         logit *= shared_dropout_layer((n_samples, options['dim_word']), use_noise, trng, retain_probability_hidden)
 
-    logit = get_layer('ff')[1](tparams, logit, options,
+    logit = get_layer_constr('ff')(tparams, logit, options,
                                prefix='ff_logit', activ='linear')
     logit_shp = logit.shape
     probs = tensor.nnet.softmax(logit.reshape([logit_shp[0]*logit_shp[1],
@@ -801,12 +336,12 @@ def build_sampler(tparams, options, use_noise, trng, return_alignment=False):
         ctx_dropout_d = theano.shared(numpy.array([1.]*4, dtype='float32'))
 
     # encoder
-    proj = get_layer(options['encoder'])[1](tparams, emb, options,
-                                            prefix='encoder', emb_dropout=emb_dropout, rec_dropout=rec_dropout)
+    proj = get_layer_constr(options['encoder'])(tparams, emb, options,
+                                            prefix='encoder', emb_dropout=emb_dropout, rec_dropout=rec_dropout, profile=profile)
 
 
-    projr = get_layer(options['encoder'])[1](tparams, embr, options,
-                                             prefix='encoder_r', emb_dropout=emb_dropout_r, rec_dropout=rec_dropout_r)
+    projr = get_layer_constr(options['encoder'])(tparams, embr, options,
+                                             prefix='encoder_r', emb_dropout=emb_dropout_r, rec_dropout=rec_dropout_r, profile=profile)
 
     # concatenate forward and backward rnn hidden states
     ctx = concatenate([proj[0], projr[0][::-1]], axis=proj[0].ndim-1)
@@ -818,7 +353,7 @@ def build_sampler(tparams, options, use_noise, trng, return_alignment=False):
     if options['use_dropout']:
         ctx_mean *= retain_probability_hidden
 
-    init_state = get_layer('ff')[1](tparams, ctx_mean, options,
+    init_state = get_layer_constr('ff')(tparams, ctx_mean, options,
                                     prefix='ff_state', activ='tanh')
 
     print >>sys.stderr, 'Building f_init...',
@@ -839,14 +374,15 @@ def build_sampler(tparams, options, use_noise, trng, return_alignment=False):
         emb *= target_dropout
 
     # apply one step of conditional gru with attention
-    proj = get_layer(options['decoder'])[1](tparams, emb, options,
+    proj = get_layer_constr(options['decoder'])(tparams, emb, options,
                                             prefix='decoder',
                                             mask=None, context=ctx,
                                             one_step=True,
                                             init_state=init_state,
                                             emb_dropout=emb_dropout_d,
                                             ctx_dropout=ctx_dropout_d,
-                                            rec_dropout=rec_dropout_d)
+                                            rec_dropout=rec_dropout_d,
+                                            profile=profile)
     # get the next hidden state
     next_state = proj[0]
 
@@ -863,18 +399,18 @@ def build_sampler(tparams, options, use_noise, trng, return_alignment=False):
     else:
         next_state_up = next_state
 
-    logit_lstm = get_layer('ff')[1](tparams, next_state_up, options,
+    logit_lstm = get_layer_constr('ff')(tparams, next_state_up, options,
                                     prefix='ff_logit_lstm', activ='linear')
-    logit_prev = get_layer('ff')[1](tparams, emb, options,
+    logit_prev = get_layer_constr('ff')(tparams, emb, options,
                                     prefix='ff_logit_prev', activ='linear')
-    logit_ctx = get_layer('ff')[1](tparams, ctxs, options,
+    logit_ctx = get_layer_constr('ff')(tparams, ctxs, options,
                                    prefix='ff_logit_ctx', activ='linear')
     logit = tensor.tanh(logit_lstm+logit_prev+logit_ctx)
 
     if options['use_dropout']:
         logit *= retain_probability_hidden
 
-    logit = get_layer('ff')[1](tparams, logit, options,
+    logit = get_layer_constr('ff')(tparams, logit, options,
                               prefix='ff_logit', activ='linear')
 
     # compute the softmax probability
@@ -1098,117 +634,6 @@ def pred_probs(f_log_probs, prepare_data, options, iterator, verbose=True, norma
     return numpy.array(probs), alignments_json
 
 
-# optimizers
-# name(hyperp, tparams, grads, inputs (list), cost) = f_grad_shared, f_update
-def adam(lr, tparams, grads, inp, cost, beta1=0.9, beta2=0.999, e=1e-8):
-
-    gshared = [theano.shared(p.get_value() * 0., name='%s_grad' % k)
-               for k, p in tparams.iteritems()]
-    gsup = [(gs, g) for gs, g in zip(gshared, grads)]
-
-    f_grad_shared = theano.function(inp, cost, updates=gsup, profile=profile)
-
-    updates = []
-
-    t_prev = theano.shared(numpy.float32(0.))
-    t = t_prev + 1.
-    lr_t = lr * tensor.sqrt(1. - beta2**t) / (1. - beta1**t)
-
-    for p, g in zip(tparams.values(), gshared):
-        m = theano.shared(p.get_value() * 0., p.name + '_mean')
-        v = theano.shared(p.get_value() * 0., p.name + '_variance')
-        m_t = beta1 * m + (1. - beta1) * g
-        v_t = beta2 * v + (1. - beta2) * g**2
-        step = lr_t * m_t / (tensor.sqrt(v_t) + e)
-        p_t = p - step
-        updates.append((m, m_t))
-        updates.append((v, v_t))
-        updates.append((p, p_t))
-    updates.append((t_prev, t))
-
-    f_update = theano.function([lr], [], updates=updates,
-                               on_unused_input='ignore', profile=profile)
-
-    return f_grad_shared, f_update
-
-def adadelta(lr, tparams, grads, inp, cost):
-    zipped_grads = [theano.shared(p.get_value() * numpy.float32(0.),
-                                  name='%s_grad' % k)
-                    for k, p in tparams.iteritems()]
-    running_up2 = [theano.shared(p.get_value() * numpy.float32(0.),
-                                 name='%s_rup2' % k)
-                   for k, p in tparams.iteritems()]
-    running_grads2 = [theano.shared(p.get_value() * numpy.float32(0.),
-                                    name='%s_rgrad2' % k)
-                      for k, p in tparams.iteritems()]
-
-    zgup = [(zg, g) for zg, g in zip(zipped_grads, grads)]
-    rg2up = [(rg2, 0.95 * rg2 + 0.05 * (g ** 2))
-             for rg2, g in zip(running_grads2, grads)]
-
-    f_grad_shared = theano.function(inp, cost, updates=zgup+rg2up,
-                                    profile=profile)
-
-    updir = [-tensor.sqrt(ru2 + 1e-6) / tensor.sqrt(rg2 + 1e-6) * zg
-             for zg, ru2, rg2 in zip(zipped_grads, running_up2,
-                                     running_grads2)]
-    ru2up = [(ru2, 0.95 * ru2 + 0.05 * (ud ** 2))
-             for ru2, ud in zip(running_up2, updir)]
-    param_up = [(p, p + ud) for p, ud in zip(itemlist(tparams), updir)]
-
-    f_update = theano.function([lr], [], updates=ru2up+param_up,
-                               on_unused_input='ignore', profile=profile)
-
-    return f_grad_shared, f_update
-
-
-def rmsprop(lr, tparams, grads, inp, cost):
-    zipped_grads = [theano.shared(p.get_value() * numpy.float32(0.),
-                                  name='%s_grad' % k)
-                    for k, p in tparams.iteritems()]
-    running_grads = [theano.shared(p.get_value() * numpy.float32(0.),
-                                   name='%s_rgrad' % k)
-                     for k, p in tparams.iteritems()]
-    running_grads2 = [theano.shared(p.get_value() * numpy.float32(0.),
-                                    name='%s_rgrad2' % k)
-                      for k, p in tparams.iteritems()]
-
-    zgup = [(zg, g) for zg, g in zip(zipped_grads, grads)]
-    rgup = [(rg, 0.95 * rg + 0.05 * g) for rg, g in zip(running_grads, grads)]
-    rg2up = [(rg2, 0.95 * rg2 + 0.05 * (g ** 2))
-             for rg2, g in zip(running_grads2, grads)]
-
-    f_grad_shared = theano.function(inp, cost, updates=zgup+rgup+rg2up,
-                                    profile=profile)
-
-    updir = [theano.shared(p.get_value() * numpy.float32(0.),
-                           name='%s_updir' % k)
-             for k, p in tparams.iteritems()]
-    updir_new = [(ud, 0.9 * ud - 1e-4 * zg / tensor.sqrt(rg2 - rg ** 2 + 1e-4))
-                 for ud, zg, rg, rg2 in zip(updir, zipped_grads, running_grads,
-                                            running_grads2)]
-    param_up = [(p, p + udn[1])
-                for p, udn in zip(itemlist(tparams), updir_new)]
-    f_update = theano.function([lr], [], updates=updir_new+param_up,
-                               on_unused_input='ignore', profile=profile)
-
-    return f_grad_shared, f_update
-
-
-def sgd(lr, tparams, grads, inp, cost):
-    gshared = [theano.shared(p.get_value() * 0.,
-                             name='%s_grad' % k)
-               for k, p in tparams.iteritems()]
-    gsup = [(gs, g) for gs, g in zip(gshared, grads)]
-
-    f_grad_shared = theano.function(inp, cost, updates=gsup,
-                                    profile=profile)
-
-    pup = [(p, p - lr * g) for p, g in zip(itemlist(tparams), gshared)]
-    f_update = theano.function([lr], [], updates=pup, profile=profile)
-
-    return f_grad_shared, f_update
-
 def train(dim_word=100,  # word vector dimensionality
           dim=1000,  # the number of LSTM units
           factors=1, # input factors
@@ -1340,7 +765,7 @@ def train(dim_word=100,  # word vector dimensionality
         print 'Reloading model parameters'
         params = load_params(saveto, params)
 
-    tparams = init_tparams(params)
+    tparams = init_theano_params(params)
 
     trng, use_noise, \
         x, x_mask, y, y_mask, \
@@ -1424,7 +849,7 @@ def train(dim_word=100,  # word vector dimensionality
     lr = tensor.scalar(name='lr')
 
     print 'Building optimizers...',
-    f_grad_shared, f_update = eval(optimizer)(lr, updated_params, grads, inps, cost)
+    f_grad_shared, f_update = eval(optimizer)(lr, updated_params, grads, inps, cost, profile=profile)
     print 'Done'
 
     print 'Optimization'
@@ -1497,7 +922,7 @@ def train(dim_word=100,  # word vector dimensionality
                 if best_p is not None:
                     params = best_p
                 else:
-                    params = unzip(tparams)
+                    params = unzip_from_theano(tparams)
                 numpy.savez(saveto, history_errs=history_errs, uidx=uidx, **params)
                 json.dump(model_options, open('%s.json' % saveto, 'wb'), indent=2)
                 print 'Done'
@@ -1508,7 +933,7 @@ def train(dim_word=100,  # word vector dimensionality
                     saveto_uidx = '{}.iter{}.npz'.format(
                         os.path.splitext(saveto)[0], uidx)
                     numpy.savez(saveto_uidx, history_errs=history_errs,
-                                uidx=uidx, **unzip(tparams))
+                                uidx=uidx, **unzip_from_theano(tparams))
                     print 'Done'
 
 
@@ -1572,7 +997,7 @@ def train(dim_word=100,  # word vector dimensionality
                 history_errs.append(valid_err)
 
                 if uidx == 0 or valid_err <= numpy.array(history_errs).min():
-                    best_p = unzip(tparams)
+                    best_p = unzip_from_theano(tparams)
                     bad_counter = 0
                 if len(history_errs) > patience and valid_err >= \
                         numpy.array(history_errs)[:-patience].min():
@@ -1583,7 +1008,7 @@ def train(dim_word=100,  # word vector dimensionality
                             print 'No progress on the validation set, increasing domain interpolation rate to %s and resuming from best params' % domain_interpolation_cur
                             train.adjust_domain_interpolation_rate(domain_interpolation_cur)
                             if best_p is not None:
-                                zipp(best_p, tparams)
+                                zip_to_theano(best_p, tparams)
                             bad_counter = 0
                         else:
                             print 'Early Stop!'
@@ -1598,7 +1023,7 @@ def train(dim_word=100,  # word vector dimensionality
                 if external_validation_script:
                     print "Calling external validation script"
                     print 'Saving  model...',
-                    params = unzip(tparams)
+                    params = unzip_from_theano(tparams)
                     numpy.savez(saveto +'.dev', history_errs=history_errs, uidx=uidx, **params)
                     json.dump(model_options, open('%s.dev.npz.json' % saveto, 'wb'), indent=2)
                     print 'Done'
@@ -1616,7 +1041,7 @@ def train(dim_word=100,  # word vector dimensionality
             break
 
     if best_p is not None:
-        zipp(best_p, tparams)
+        zip_to_theano(best_p, tparams)
 
     use_noise.set_value(0.)
     valid_errs, alignment = pred_probs(f_log_probs, prepare_data,
@@ -1628,7 +1053,7 @@ def train(dim_word=100,  # word vector dimensionality
     if best_p is not None:
         params = copy.copy(best_p)
     else:
-        params = unzip(tparams)
+        params = unzip_from_theano(tparams)
     numpy.savez(saveto, zipped_params=best_p,
                 history_errs=history_errs,
                 uidx=uidx,
